abstracted search function from search component
