+++ packages/lms/src/routes/api/search/+server.ts
@@ -1,10 +1,9 @@
 import { graphql } from 'graphql';
-import { schema, resolvers } from '$lib/graphql/graphql-schema';
+import { schema, resolvers } from '$lib/utils/graphql/graphQLSchema';
 import { parseMarkdownSearch } from '$lib/utils/parseMarkdownSearch';
 
 // Pass the data from the parseMarkdown function to the GraphQL context
 const data = parseMarkdownSearch('src/routes/content');
-console.log('data:', data);
 
 export async function POST({ request }) {
 	const { query, variables } = await request.json();
